# PoC: Replace `zlib` with `zlib-ng`