--- a/cmake/targets/BuildLibArchive.cmake
+++ b/cmake/targets/BuildLibArchive.cmake
@@ -38,7 +38,7 @@ register_cmake_command(
     # spawn a processes to compress instead of using the library.
     -DENABLE_ZLIB=OFF
     -DHAVE_ZLIB_H=ON
-    -DCMAKE_C_FLAGS="-I${VENDOR_PATH}/zlib"
+    -DCMAKE_C_FLAGS="-I${BUILD_PATH}/zlib"
   LIB_PATH
     libarchive
   LIBRARIES