… cmake/Globals.cmake
… patches/zlib/deflate.h.patch