### What does this PR do?

It was mentioned a while ago on the official Discord server about replacing the Cloudflare zlib fork with the zlib-ng one, as it already includes the changes from the CF fork, architectural and QoL improvements over the original project (the full manifesto explaining the ["why"](https://github.com/zlib-ng/zlib-ng?tab=readme-ov-file#history) of this fork)

This is just a PoC for now.

- [ ] Documentation or TypeScript types (it's okay to leave the rest blank in this case)
- [x] Code changes

### How did you verify your code works?

CI (?)