@inetol 2025-01-01
I have not seen weird things or crashes, at least on my Linux workstation. I don't know the rest of OS if there are any, please approve the workflows to better test this PR