@inetol 2025-01-05
> Looks like it does not compile due to missing includes.

Should be fixed now

> my guess is that inflate streaming performance is regressed a little compared to cloudflare-zlib

IDK if it has anything to do with zlib compat layer, although the cleanest thing to do would be to refactor zlib to use the new API