@Gobd 2025-02-25
I saw this posted on HN not sure if that's something you'd be interested in trying. Says zlib-rs is even faster than zlib-ng https://news.ycombinator.com/item?id=43171413