@inetol 2025-02-25
I also saw something a while ago, but it *doesn't look like* they are at 100% parity in functionality like zlib-ng does rn. Anyway, if someone would like to give it a try, they can always open another PR