@inetol 2025-03-25
@Jarred-Sumner It seems that the regression with Next.js has been solved, running the benchmarks in this last stable rebase..:

**v1.2.6:** 5184/rps
**e34b7e3601379271504a6a501fe0e07784a7e5d5:** 5096/rps

As I'm running it directly from my workstation there may be inconsistencies, but no longer shows the clear performance penalty of last time.